Removing trailing newline for safer editing
