--- repo/common.go
@@ -1,25 +1,0 @@
-package repo
-
-import "fmt"
-
-const (
-	actionCreate = iota
-	actionUpdate
-	actionDelete
-)
-
-var commitMsg = map[int]string{
-	actionCreate: "Create",
-	actionUpdate: "Update",
-	actionDelete: "Delete",
-}
-
-func strPtr(s string) *string { return &s }
-
-func repoAddress(owner, name string) string {
-	return fmt.Sprintf("https://github.com/%s/%s", owner, name)
-}
-
-func uploadAddress(owner, name, file string) string {
-	return fmt.Sprintf("https://api.github.com/repos/%s/%s/contents/%s", owner, name, file)
-}
